Set navigation to manage expense page on pressing either on header plus button or on one of the expense items
If a component is not registered as a screen it cannot accept a navigation property. In this case the useNavigation hook can help

And for the add button in the tab navigator the options object can be modified by using it as a function returning these options. This function can then accept the destructured built-in navigation prop